refactor: 중복되는 테스트 Fixture 제거
