--- src/test/intg/java/com/bob/support/fixture/auth/MailFixture.java
@@ -1,8 +1,0 @@
-package com.bob.support.fixture.auth;
-
-public class MailFixture {
-
-  public static final String EMAIL = "test@email.com";
-  public static final String VALID_CODE = "ABC123";
-  public static final String INVALID_CODE = "XYZ789";
-}
