--- src/test/intg/java/com/bob/support/fixture/command/CreateMemberCommandFixture.java
@@ -1,9 +1,0 @@
-package com.bob.support.fixture.command;
-
-import com.bob.domain.member.command.CreateMemberCommand;
-
-public class CreateMemberCommandFixture {
-  public static CreateMemberCommand defaultCreateMemberCommand() {
-    return new CreateMemberCommand("test@email.com", "password", "tester", 213);
-  }
-}
